v1 tests failing, concept for splitting class
