+++ lib/debt_ceiling.rb
@@ -4,8 +4,9 @@ require 'rubycritic'
 require 'rubycritic/cli/application'
 require 'ostruct'
 require_relative 'debt_ceiling/accounting'
-require_relative 'debt_ceiling/custom_debt_analysis'
 require_relative 'debt_ceiling/debt'
+require_relative 'debt_ceiling/static_analysis_debt'
+require_relative 'debt_ceiling/explicit_assigned_debt'
 require_relative 'debt_ceiling/compatibility'
 require_relative 'debt_ceiling/file_attributes'
