+++ lib/debt_ceiling/debt.rb
@@ -2,13 +2,8 @@ require 'forwardable'
 module DebtCeiling
   class Debt
     extend Forwardable
-    include CustomDebtAnalysis
     DoNotWhitelistAndBlacklistSimulateneously = Class.new(StandardError)
 
-    def_delegators :file_attributes, :path, :analysed_module, :module_name, :linecount, :source_code
-    def_delegators :non_grade_scoring, :complexity_multiplier, :duplication_multiplier, :smells_multiplier,
-                   :method_count_multiplier, :ideal_max_line_count, :cost_per_line_over_ideal
-    def_delegator :analysed_module, :rating
     attr_accessor :debt_amount
     def_delegator :debt_amount, :to_i
 
@@ -37,40 +32,12 @@ module DebtCeiling
       self.debt_amount = external_measure_debt || internal_measure_debt
     end
 
-    def internal_measure_debt
-      external_augmented_debt +
-      cost_from_static_analysis_points +
-      debt_from_source_code_rules
+    def external_measure_debt
+      public_send(:measure_debt) if self.respond_to?(:measure_debt)
     end
 
-    def cost_from_static_analysis_points
-      DebtCeiling.grade_points[letter_grade] + cost_from_non_grade_scoring
-    end
-
-    def cost_from_non_grade_scoring
-      flog_flay_debt + method_count_debt + smells_debt + line_count_debt
-    end
-
-    def smells_debt
-      analysed_module.smells.map(&:cost).inject(0, :+) * smells_multiplier
-    end
-
-    def method_count_debt
-      analysed_module.methods_count * method_count_multiplier
-    end
-
-    def flog_flay_debt
-      analysed_module.complexity *  complexity_multiplier +
-      analysed_module.duplication * duplication_multiplier
-    end
-
-    def line_count_debt
-      excess_lines = linecount - ideal_max_line_count
-      excess_lines > 0 ? excess_lines * cost_per_line_over_ideal : 0
-    end
-
-    def non_grade_scoring
-      DebtCeiling
+    def external_augmented_debt
+      (public_send(:augment_debt) if respond_to?(:augment_debt)).to_i
     end
 
     def self.whitelist_includes?(debt)
@@ -80,5 +47,15 @@ module DebtCeiling
     def self.blacklist_includes?(debt)
       DebtCeiling.blacklist.find { |filename| debt.path.match filename }
     end
+
+    def valid_debt?
+      black_empty = DebtCeiling.blacklist.empty?
+      white_empty = DebtCeiling.whitelist.empty?
+      fail DoNotWhitelistAndBlacklistSimulateneously unless black_empty || white_empty
+      (black_empty && white_empty) ||
+      (black_empty && self.class.whitelist_includes?(self)) ||
+      (white_empty && !self.class.blacklist_includes?(self))
+    end
+
   end
 end
